requirements.txt
requests==2.10.0
Cerberus==1.2
PyGithub==1.39
ruamel.yaml==0.15.37
ruamel.ordereddict==0.4.13